Separate Organization seed file, add address cleanup to seed.
